# api/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Grocery & Meal Planner API
  description: |
    MVP API for managing pantry inventory, grocery lists, meal ideas, and expiration notifications.
    
    ## Features
    - **Inventory Management**: CRUD operations for pantry items with low-stock and expiring-soon flags
    - **Grocery Lists**: Create and manage shopping lists, auto-add low stock or expiring items
    - **Meal Ideas**: Get suggestions based on pantry contents, save favorite meals
    - **Notifications**: Schedule reminders for expiring items
  version: 1.0.0
  contact:
    name: Smart Grocery Team
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/grocery-lists:
    get:
      tags: [Grocery Lists]
      summary: List grocery lists
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, archived]
      responses:
        '200':
          description: List of grocery lists
    post:
      tags: [Grocery Lists]
      summary: Create grocery list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroceryList'
      responses:
        '201':
          description: List created
  /api/grocery-lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Grocery Lists]
      summary: Get grocery list with items
      responses:
        '200':
          description: Grocery list
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Grocery Lists]
      summary: Update grocery list
      responses:
        '200':
          description: List updated
    delete:
      tags: [Grocery Lists]
      summary: Delete grocery list
      responses:
        '200':
          description: List deleted
  /api/grocery-lists/{id}/items:
    post:
      tags: [Grocery Lists]
      summary: Add item to list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroceryListItem'
      responses:
        '201':
          description: Item added
  /api/grocery-lists/{listId}/items/{itemId}:
    parameters:
      - name: listId
        in: path
        required: true
        schema:
          type: integer
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [Grocery Lists]
      summary: Update list item (check/uncheck)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroceryListItem'
      responses:
        '200':
          description: Item updated
    delete:
      tags: [Grocery Lists]
      summary: Remove item from list
      responses:
        '200':
          description: Item removed
  /api/grocery-lists/{id}/add-expiring:
    post:
      tags: [Grocery Lists]
      summary: Auto-add expiring pantry items to list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Items added
  /api/grocery-lists/{id}/add-low-stock:
    post:
      tags: [Grocery Lists]
      summary: Auto-add low stock items to list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Items added

  # Meals
...
components:
  schemas:
...
    CreateGroceryList:
      type: object
      required: [userId, title]
      properties:
        userId:
          type: integer
        title:
          type: string
          maxLength: 255
        status:
          type: string
          enum: [active, completed, archived]
          default: active
    CreateGroceryListItem:
      type: object
      required: [name]
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
        note:
          type: string
        pantryItemId:
          type: integer
    UpdateGroceryListItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        note:
          type: string
        isChecked:
          type: boolean

    # Meal schemas
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  message:
                    type: string
...